pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semanrag"
version = "0.1.0"
description = "Semantic RAG — knowledge-graph-augmented retrieval-augmented generation"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{name = "SemanRAG Contributors"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "tiktoken>=0.5.0",
    "networkx>=3.1",
    "graspologic>=3.3.0",
    "nano-vectordb>=0.0.5",
    "rank-bm25>=0.2.2",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "rapidfuzz>=3.5.0",
    "json_repair>=0.28.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "xlsxwriter>=3.1.0",
    "pypinyin>=0.49.0",
    "pipmaster>=0.1.0",
    "configparser>=6.0.0",
    "google-genai>=1.0.0",
    "google-api-core>=2.15.0",
    "packaging>=23.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "gunicorn>=21.2.0",
    "PyJWT>=2.8.0",
    "python-jose>=3.3.0",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "slowapi>=0.1.9",
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
]
multimodal = [
    "camelot-py>=0.11.0",
    "docling>=2.0.0",
    "pytesseract>=0.3.10",
    "rapidocr-onnxruntime>=1.3.0",
]
offline-storage = [
    "redis>=5.0.0",
    "neo4j>=5.14.0",
    "pymilvus>=2.3.0",
    "pymongo>=4.6.0",
    "motor>=3.3.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.4",
    "qdrant-client>=1.7.0",
    "opensearch-py>=2.4.0",
    "tantivy>=0.21.0",
]
offline-llm = [
    "openai>=1.6.0",
    "anthropic>=0.18.0",
    "ollama>=0.1.6",
    "zhipuai>=2.0.0",
    "aioboto3>=12.0.0",
    "google-genai>=1.0.0",
    "llama-index>=0.10.0",
    "sentence-transformers>=2.3.0",
    "FlagEmbedding>=1.2.0",
]
safety = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
]
queue = [
    "celery>=5.3.0",
    "arq>=0.25.0",
]
observability = [
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "prometheus-client>=0.19.0",
    "langfuse>=2.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
evaluation = [
    "ragas>=0.1.0",
    "datasets>=2.16.0",
]
export = [
    "rdflib>=7.0.0",
]

[project.scripts]
semanrag = "semanrag.cli:main"
semanrag-server = "semanrag.api.semanrag_server:main"
semanrag-gunicorn = "semanrag.api.semanrag_server:gunicorn_main"
semanrag-hash-password = "semanrag.tools.hash_password:main"
semanrag-download-cache = "semanrag.tools.download_cache:main"
semanrag-clean-llmqc = "semanrag.tools.clean_llm_query_cache:main"
semanrag-check-init = "semanrag.tools.check_initialization:main"
semanrag-scan-orphans = "semanrag.tools.scan_orphans:main"
semanrag-scan-stale = "semanrag.tools.scan_stale:main"
semanrag-resolve-entities = "semanrag.tools.resolve_entities:main"
semanrag-migrate-cache = "semanrag.tools.migrate_llm_cache:main"
semanrag-eval = "semanrag.evaluation.runner:main"

[tool.setuptools.packages.find]
include = ["semanrag*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: fast unit tests",
    "offline: tests that do not require network access",
    "integration: end-to-end integration tests",
    "requires_db: tests that need a running database",
    "requires_api: tests that need a running API server",
    "requires_vision: tests that need vision model access",
    "regression: regression tests for known issues",
]
